+++ docs/src/pages/[platform]/components/passwordfield/examples/PasswordFieldIconExample.tsx
@@ -1,0 +1,15 @@
+import { PasswordField, IconsProvider } from '@aws-amplify/ui-react';
+import { FiLock, FiUnlock } from 'react-icons/fi';
+
+export const PasswordFieldIconExample = () => (
+  <IconsProvider
+    icons={{
+      passwordField: {
+        visibility: <FiLock />,
+        visibilityOff: <FiUnlock />,
+      },
+    }}
+  >
+    <PasswordField label="Password" />
+  </IconsProvider>
+);