+++ packages/react/src/primitives/Icon/context/IconsProvider.tsx
@@ -1,0 +1,18 @@
+import * as React from 'react';
+import { IconsContext, IconsContextInterface } from './IconsContext';
+
+interface IconProviderProps {
+  children: React.ReactNode;
+  icons?: IconsContextInterface;
+}
+
+export { IconsContextInterface };
+
+export function IconsProvider({
+  children,
+  icons,
+}: IconProviderProps): JSX.Element {
+  return (
+    <IconsContext.Provider value={icons}>{children}</IconsContext.Provider>
+  );
+}