+++ packages/react-notifications/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/ui-react-notifications
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`5cc76794a`](https://github.com/aws-amplify/amplify-ui/commit/5cc76794a71584e26b1ec699c1dc1713d8a986c9), [`f8c5b77e4`](https://github.com/aws-amplify/amplify-ui/commit/f8c5b77e46f2b304e33776b28ed5a18141f9b364), [`00f2344e5`](https://github.com/aws-amplify/amplify-ui/commit/00f2344e593b47d6a7413869220606f32e5eaf1a)]:
+  - @aws-amplify/ui@5.7.2
+  - @aws-amplify/ui-react@5.2.0
+  - @aws-amplify/ui-react-core-notifications@1.0.8
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`00c7abc25`](https://github.com/aws-amplify/amplify-ui/commit/00c7abc25b263bb42b67d8980b31212d0a206d1c), [`54d884dd0`](https://github.com/aws-amplify/amplify-ui/commit/54d884dd0ae7f16cc1f5b71ae767e0ccf477c4b5), [`5040faf51`](https://github.com/aws-amplify/amplify-ui/commit/5040faf51ce2dc87882d452e6f90ad4ab0bd6967)]:
+  - @aws-amplify/ui-react@5.1.1
+  - @aws-amplify/ui@5.7.1
+  - @aws-amplify/ui-react-core-notifications@1.0.7
+
 ## 1.0.9
 
 ### Patch Changes